Changes after review.
